+++ src/commands/mod.rs
@@ -1,5 +1,7 @@
 pub mod clean;
 pub mod clean_all;
+pub mod config;
+pub mod info;
 pub mod list;
 pub mod run;
 pub mod setup;