+++ src/utils/mod.rs
@@ -1,3 +1,5 @@
+pub mod env;
 pub mod git;
 pub mod path;
 pub mod process;
+pub mod shell;